Changing email_id to bigint

+++ database/migrations/2016_06_29_154724_create_attribution_record_truths_table.php
@@ -16,7 +16,7 @@ class CreateAttributionRecordTruthsTable extends Migration
     public function up()
     {
         Schema::create('attribution_record_truths', function (Blueprint $table) {
-            $table->integer( 'email_id' )->unsigned();
+            $table->bigInteger( 'email_id' )->unsigned();
             $table->boolean( 'expired' )->default( false );
             $table->boolean( 'active' )->default( false );
             $table->timestamps();
